tests: add end-to-end amendment tests for Btk and Ptk
With yaml instead of json, because its prettier.

Most semantic data is stripped from the files to have smaller test files,
and they are not really needed for the code to work. They are faked back,
or else the first semantic reparse would take a ton of time during the
amending phase.