web: port hun_law html writer
But without the Etree stuff. That's not as elegant I thought it would be.
The code is way shorter this way.

Also ported the style.css as-is. It should be completely rewritten though,
because the huge grid layout is broken in chrome.

And created a nice little list view as index for now.